Use 64-bit index for JPEG coefficient blocks to fix large-image lossless reconstruction

+++ lib/jxl/jpeg/enc_jpeg_data_reader.cc
@@ -833,7 +833,8 @@ Status ProcessScan(const uint8_t* data, const size_t len,
           for (int ix = 0; ix < nblocks_x; ++ix) {
             int block_y = mcu_y * nblocks_y + iy;
             int block_x = mcu_x * nblocks_x + ix;
-            int block_idx = block_y * c->width_in_blocks + block_x;
+            size_t block_idx =
+                static_cast<size_t>(block_y) * c->width_in_blocks + block_x;
             bool reset_state = false;
             int num_zero_runs = 0;
             coeff_t* coeffs = &c->coeffs[block_idx * kDCTBlockSize];
